+++ frontend/css/_FeedbackModal.scss
@@ -87,6 +87,62 @@
     flex-direction: column;
 
   }
+  &__initialText {
+    margin-top: 10px;
+    padding-left: 32px;
+    margin-left: 12px;
+    flex-grow: 1;
+    cursor: pointer;
+    user-select: none;
+    font-size: 12px;
+    position: relative;
+    width: 200px;
+    
+    &:hover {
+      background: #eaeaea;
+    }
+   
+  }
+
+  &__initialText > input {
+    position: absolute;
+    opacity: 0;
+    cursor: pointer;
+    height: 16px;
+    width: 16px;
+  }
+
+  &__initialText > input:checked + .FeedbackModal__initialRadio::after {
+    display: block;
+
+  }
+
+  &__initialRadio {
+    position: absolute;
+    top: 4px;
+    left: 8px;
+    height: 14px;
+    width: 14px;
+    background-color: transparent;
+    border-radius: 50%;
+    border: 1px solid rgba(0, 0, 0, 0.42);
+
+    &:hover {
+      cursor: pointer;
+    }
+  }
+
+  &__initialRadio:after {
+    content: "";
+    position: absolute;
+    display: none;
+    top: 2px;
+    left: 2px;
+    width: 8px;
+    height: 8px;
+    border-radius: 50%;
+    background: #338ef1;
+  }
 
   &__checkBoxes {
     width: 200px;
